+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Commander Changelog
 
+## unreleased
+
+### Enhancements
+
 - Renamed VaradicArgument to VariadicArgument.
+- Adds support for variadic options, allowing the user to repeat options to
+  provide additional values.
 
 ## 0.5.0